refactored App.js, moving getLoadingLabel into separate file
